{{cookiecutter.project_slug}}/requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.11.1
inflection==0.5.1
jsonschema==3.2.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyrsistent==0.17.3
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
django-cors-headers
requests
celery==5.0.5
redis==3.5.3
